Refactor to folder
